chore: remove kyc and balance details from tabs groups
