+++ apps/admin-panel/app/customers/[customer-id]/page.stories.tsx
@@ -1,25 +1,26 @@
 import type { Meta, StoryObj } from "@storybook/react"
 import { MockedProvider } from "@apollo/client/testing"
 
-import CustomerPage from "./page"
 import CustomerLayout from "./layout"
 
+import CustomerTransactionsPage from "./page"
+
 import {
-  GetCustomerOverviewDocument,
+  GetCustomerTransactionsDocument,
   GetCustomerBasicDetailsDocument,
   AccountStatus,
 } from "@/lib/graphql/generated"
 
 const meta = {
-  title: "Pages/Customers/Customer/Overview",
-  component: CustomerPage,
+  title: "Pages/Customers/Customer/Transactions",
+  component: CustomerTransactionsPage,
   parameters: {
     layout: "fullscreen",
     nextjs: {
       appDirectory: true,
     },
   },
-} satisfies Meta<typeof CustomerPage>
+} satisfies Meta<typeof CustomerTransactionsPage>
 
 export default meta
 type Story = StoryObj<typeof meta>
@@ -50,10 +51,10 @@ const layoutMocks = [
   },
 ]
 
-const overviewMocks = [
+const transactionsMocks = [
   {
     request: {
-      query: GetCustomerOverviewDocument,
+      query: GetCustomerTransactionsDocument,
       variables: {
         id: "4178b451-c9cb-4841-b248-5cc20e7774a6",
       },
@@ -62,49 +63,57 @@ const overviewMocks = [
       data: {
         customer: {
           id: "Customer:4178b451-c9cb-4841-b248-5cc20e7774a6",
-          customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
-          balance: {
-            checking: {
-              settled: 1000,
-              pending: 500,
+          deposits: [
+            {
+              createdAt: "2024-11-25T06:25:30.866119Z",
+              customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+              depositId: "dep-123",
+              reference: "DEP123",
+              amount: 1000,
             },
-          },
+          ],
+          withdrawals: [
+            {
+              status: "COMPLETED",
+              reference: "WIT123",
+              customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+              withdrawalId: "wit-123",
+              createdAt: "2024-11-25T06:25:30.866119Z",
+              amount: 500,
+              customer: {
+                customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+                email: "test@lana.com",
+              },
+            },
+          ],
+          transactions: [
+            {
+              __typename: "Deposit",
+              createdAt: "2024-11-25T06:25:30.866119Z",
+              customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+              depositId: "dep-123",
+              reference: "DEP123",
+              amount: 1000,
+            },
+            {
+              __typename: "Withdrawal",
+              status: "COMPLETED",
+              reference: "WIT123",
+              customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+              withdrawalId: "wit-123",
+              createdAt: "2024-11-25T06:25:30.866119Z",
+              amount: 500,
+              customer: {
+                customerId: "4178b451-c9cb-4841-b248-5cc20e7774a6",
+                email: "test@lana.com",
+              },
+            },
+          ],
         },
       },
     },
   },
 ]
-
-const LoadingStory = () => {
-  const mocks = [
-    {
-      request: {
-        query: GetCustomerBasicDetailsDocument,
-        variables: {
-          id: "4178b451-c9cb-4841-b248-5cc20e7774a6",
-        },
-      },
-      delay: Infinity,
-    },
-    {
-      request: {
-        query: GetCustomerOverviewDocument,
-        variables: {
-          id: "4178b451-c9cb-4841-b248-5cc20e7774a6",
-        },
-      },
-      delay: Infinity,
-    },
-  ]
-
-  return (
-    <MockedProvider mocks={mocks} addTypename={false}>
-      <CustomerLayout params={mockParams}>
-        <CustomerPage params={mockParams} />
-      </CustomerLayout>
-    </MockedProvider>
-  )
-}
 
 export const Default: Story = {
   args: {
@@ -114,25 +123,11 @@ export const Default: Story = {
     (Story) => (
       <MockedProvider mocks={layoutMocks} addTypename={false}>
         <CustomerLayout params={mockParams}>
-          <MockedProvider mocks={overviewMocks} addTypename={false}>
+          <MockedProvider mocks={transactionsMocks} addTypename={false}>
             <Story />
           </MockedProvider>
         </CustomerLayout>
       </MockedProvider>
     ),
   ],
-}
-
-export const Loading: Story = {
-  args: {
-    params: mockParams,
-  },
-  render: LoadingStory,
-  parameters: {
-    nextjs: {
-      navigation: {
-        pathname: "/customers/[customer-id]",
-      },
-    },
-  },
 }
